+++ contracts/mocks/ERC165/ERC165NotSupported.sol
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.0;
 
 contract ERC165NotSupported { }