+++ contracts/mocks/ERC777Mock.sol
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.0;
 
 import "../GSN/Context.sol";
@@ -21,5 +23,9 @@ contract ERC777Mock is Context, ERC777 {
         bytes memory operatorData
     ) public {
         _mint(to, amount, userData, operatorData);
+    }
+
+    function approveInternal(address holder, address spender, uint256 value) public {
+        _approve(holder, spender, value);
     }
 }