+++ contracts/token/ERC1155/IERC1155MetadataURI.sol
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.2;
+
+import "./IERC1155.sol";
+
+/**
+ * @dev Interface of the optional ERC1155MetadataExtension interface, as defined
+ * in the https://eips.ethereum.org/EIPS/eip-1155#metadata-extensions[EIP].
+ */
+interface IERC1155MetadataURI is IERC1155 {
+    function uri(uint256 id) external view returns (string memory);
+}