+++ contracts/token/ERC721/IERC721Enumerable.sol
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.2;
 
 import "./IERC721.sol";