+++ contracts/utils/Pausable.sol
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.0;
 
 import "../GSN/Context.sol";
@@ -40,6 +42,10 @@ contract Pausable is Context {
 
     /**
      * @dev Modifier to make a function callable only when the contract is not paused.
+     *
+     * Requirements:
+     *
+     * - The contract must not be paused.
      */
     modifier whenNotPaused() {
         require(!_paused, "Pausable: paused");
@@ -48,6 +54,10 @@ contract Pausable is Context {
 
     /**
      * @dev Modifier to make a function callable only when the contract is paused.
+     *
+     * Requirements:
+     *
+     * - The contract must be paused.
      */
     modifier whenPaused() {
         require(_paused, "Pausable: not paused");
@@ -56,6 +66,10 @@ contract Pausable is Context {
 
     /**
      * @dev Triggers stopped state.
+     *
+     * Requirements:
+     *
+     * - The contract must not be paused.
      */
     function _pause() internal virtual whenNotPaused {
         _paused = true;
@@ -64,6 +78,10 @@ contract Pausable is Context {
 
     /**
      * @dev Returns to normal state.
+     *
+     * Requirements:
+     *
+     * - The contract must be paused.
      */
     function _unpause() internal virtual whenPaused {
         _paused = false;