+++ test/helpers/merkleTree.js
@@ -1,9 +1,9 @@
-const { keccak256, bufferToHex } = require('ethereumjs-util');
+const { keccak256, keccakFromString, bufferToHex } = require('ethereumjs-util');
 
 class MerkleTree {
   constructor (elements) {
     // Filter empty strings and hash elements
-    this.elements = elements.filter(el => el).map(el => keccak256(el));
+    this.elements = elements.filter(el => el).map(el => keccakFromString(el));
 
     // Sort elements
     this.elements.sort(Buffer.compare);
@@ -97,7 +97,7 @@ class MerkleTree {
 
     // Convert element to 32 byte hash if it is not one already
     if (el.length !== 32 || !Buffer.isBuffer(el)) {
-      hash = keccak256(el);
+      hash = keccakFromString(el);
     } else {
       hash = el;
     }